to move lib/cmd/run/generator.go to lib/generator/generator.go
