+++ lib/cmd/gentx/gentx.go
@@ -3,11 +3,11 @@ package gentx
 import (
 	"log"
 
-	"github.com/0glabs/evmchainbench/lib/cmd/run"
+	generatorpkg "github.com/0glabs/evmchainbench/lib/generator"
 )
 
 func GenTx(rpcUrl, faucetPrivateKey string, senderCount, txCount int, txStoreDir string) {
-	generator, err := run.NewGenerator(rpcUrl, faucetPrivateKey, senderCount, txCount, true, txStoreDir)
+	generator, err := generatorpkg.NewGenerator(rpcUrl, faucetPrivateKey, senderCount, txCount, true, txStoreDir)
 	if err != nil {
 		log.Fatalf("Failed to create generator: %v", err)
 	}
