+++ lib/cmd/run/run.go
@@ -3,10 +3,12 @@ package run
 import (
 	"log"
 	"sync"
+
+	"github.com/0glabs/evmchainbench/lib/generator"
 )
 
 func Run(rpcUrl, faucetPrivateKey string, senderCount, txCount int) {
-	generator, err := NewGenerator(rpcUrl, faucetPrivateKey, senderCount, txCount, false, "")
+	generator, err := generator.NewGenerator(rpcUrl, faucetPrivateKey, senderCount, txCount, false, "")
 	if err != nil {
 		log.Fatalf("Failed to create generator: %v", err)
 	}
