+++ lib/generator/generator.go
@@ -1,4 +1,4 @@
-package run
+package generator
 
 import (
 	"context"
